feat(AvatarUpload): add focused state
